+++ signalfx/resource_signalfx_aws_token_integration.go
@@ -1,6 +1,7 @@
 package signalfx
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"log"
@@ -48,7 +49,7 @@ func integrationAWSTokenResource() *schema.Resource {
 
 func integrationAWSTokenExists(d *schema.ResourceData, meta interface{}) (bool, error) {
 	config := meta.(*signalfxConfig)
-	_, err := config.Client.GetAWSCloudWatchIntegration(d.Id())
+	_, err := config.Client.GetAWSCloudWatchIntegration(context.TODO(), d.Id())
 	if err != nil {
 		if strings.Contains(err.Error(), "404") {
 			return false, nil
@@ -60,7 +61,7 @@ func integrationAWSTokenExists(d *schema.ResourceData, meta interface{}) (bool, 
 
 func integrationAWSTokenRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*signalfxConfig)
-	_, err := config.Client.GetAWSCloudWatchIntegration(d.Id())
+	_, err := config.Client.GetAWSCloudWatchIntegration(context.TODO(), d.Id())
 	if err != nil {
 		if strings.Contains(err.Error(), "404") {
 			d.SetId("")
@@ -99,7 +100,7 @@ func integrationAWSTokenCreate(d *schema.ResourceData, meta interface{}) error {
 	debugOutput, _ := json.Marshal(payload)
 	log.Printf("[DEBUG] SignalFx: Create AWS Token Integration Payload: %s", string(debugOutput))
 
-	int, err := config.Client.CreateAWSCloudWatchIntegration(payload)
+	int, err := config.Client.CreateAWSCloudWatchIntegration(context.TODO(), payload)
 	if err != nil {
 		if strings.Contains(err.Error(), "40") {
 			err = fmt.Errorf("%s\nPlease verify you are using an admin token when working with integrations", err.Error())
@@ -125,5 +126,5 @@ func integrationAWSTokenCreate(d *schema.ResourceData, meta interface{}) error {
 func integrationAWSTokenDelete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*signalfxConfig)
 
-	return config.Client.DeleteAWSCloudWatchIntegration(d.Id())
+	return config.Client.DeleteAWSCloudWatchIntegration(context.TODO(), d.Id())
 }