+++ signalfx/resource_signalfx_text_chart.go
@@ -1,6 +1,7 @@
 package signalfx
 
 import (
+	"context"
 	"encoding/json"
 	"log"
 
@@ -65,7 +66,7 @@ func textchartCreate(d *schema.ResourceData, meta interface{}) error {
 	debugOutput, _ := json.Marshal(payload)
 	log.Printf("[DEBUG] SignalFx: Create Text Chart Payload: %s", string(debugOutput))
 
-	c, err := config.Client.CreateChart(payload)
+	c, err := config.Client.CreateChart(context.TODO(), payload)
 	if err != nil {
 		return err
 	}
@@ -100,7 +101,7 @@ func textchartAPIToTF(d *schema.ResourceData, c *chart.Chart) error {
 
 func textchartRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*signalfxConfig)
-	c, err := config.Client.GetChart(d.Id())
+	c, err := config.Client.GetChart(context.TODO(), d.Id())
 	if err != nil {
 		return err
 	}
@@ -122,7 +123,7 @@ func textchartUpdate(d *schema.ResourceData, meta interface{}) error {
 	debugOutput, _ := json.Marshal(payload)
 	log.Printf("[DEBUG] SignalFx: Update Text Chart Payload: %s", string(debugOutput))
 
-	c, err := config.Client.UpdateChart(d.Id(), payload)
+	c, err := config.Client.UpdateChart(context.TODO(), d.Id(), payload)
 	if err != nil {
 		return err
 	}
@@ -135,5 +136,5 @@ func textchartUpdate(d *schema.ResourceData, meta interface{}) error {
 func textchartDelete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*signalfxConfig)
 
-	return config.Client.DeleteChart(d.Id())
+	return config.Client.DeleteChart(context.TODO(), d.Id())
 }